// src/app/@core/models/rca.ts
export class RcaProject {
    id?: string;
    name?: string;
    description?: string;
    startDate?: string;
    endDate?: string;
    status?: string;
    owner?: string;
    rootCause?: string;
    correctiveAction?: string;
    category?: string;
    teamMembers?: string[];
    priority?: string;
    impact?: string;
  }